move AppServerForTesting from common to server
