+++ src/test/java/com/scmspain/karyon/healthcheckendpoint/annotation/server/AppServerForTesting.java
@@ -1,4 +1,4 @@
-package com.scmspain.karyon.healthcheckendpoint.annotation.common;
+package com.scmspain.karyon.healthcheckendpoint.annotation.server;
 
 import com.google.inject.Singleton;
 import com.netflix.governator.annotations.Modules;
